Fixes crash looking up config file / run code
Fixes #111
Re enabled pgs look and feel

+++ desktop/src/gaia/cu9/ari/gaiaorbit/desktop/GaiaSkyDesktop.java
@@ -96,14 +96,8 @@ public class GaiaSkyDesktop implements IObserver {
             }
 
             // Swing look and feel
-            // This fixes HiDPI in linux
-            if (SysUtils.isLinux()) {
-                UIManager.setLookAndFeel("com.sun.java.swing.plaf.gtk.GTKLookAndFeel");
-            } else if (SysUtils.isWindows()) {
-                UIManager.setLookAndFeel("com.sun.java.swing.plaf.windows.WindowsLookAndFeel");
-            } else {
-                UIManager.setLookAndFeel(UIManager.getSystemLookAndFeelClassName());
-            }
+            UIManager.setLookAndFeel("com.pagosoft.plaf.PgsLookAndFeel");
+            
 
             // Init global configuration
             ConfInit.initialize(new DesktopConfInit());
@@ -316,7 +310,17 @@ public class GaiaSkyDesktop implements IObserver {
 
         if (ow || !userFolderConfFile.exists()) {
             // Copy file
-            copyFile(new File("conf" + File.separator + "global.properties"), userFolderConfFile, ow);
+        	File confFolder = new File("conf" + File.separator);
+        	if(confFolder.exists() && confFolder.isDirectory()){
+        		// Running released package
+        		copyFile(new File("conf" + File.separator + "global.properties"), userFolderConfFile, ow);
+        	}else{
+        		// Running from code?
+        		if(!new File("../android/assets/conf" + File.separator).exists()){
+        			throw new IOException("File ../android/assets/conf does not exist!");
+        		}
+        		copyFile(new File("../android/assets/conf" + File.separator + "global.properties"), userFolderConfFile, ow);
+        	}
         }
         String props = userFolderConfFile.getAbsolutePath();
         System.setProperty("properties.file", props);
